convert *speedgraphval to double

+++ src/net_stat.h
@@ -71,8 +71,8 @@ void print_addrs(struct text_object *, char *, int);
 #endif /* __linux__ */
 #ifdef X11
 void parse_net_stat_graph_arg(struct text_object *, const char *, void *);
-uint8_t downspeedgraphval(struct text_object *);
-uint8_t upspeedgraphval(struct text_object *);
+double downspeedgraphval(struct text_object *);
+double upspeedgraphval(struct text_object *);
 #endif /* X11 */
 #ifdef __linux__
 #ifdef HAVE_IWLIB
